Original Shortcuts 2.1 sources from September 29, 2019
Shortcuts is an app developed for loading legacy contextual menu plug-ins in macOS after Apple removed in-process plug-in loading.
The main Shortcuts.app with GUI is used to set up the system. The background ShortcutObserver app is employed to load the plug-ins and execute their actions in a front app, for example Finder for file selection or TexEdit for text context.
The app was 64-bit but only for Intel macs.
More information at:
https://www.abracode.com/free/cmworkshop/shortcuts.html